File restructure
Schema.org receives its own module wherein strategies for extracting
attributes lie.
